+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Roadmap / Unreleased]
+- Add breadcrumb navigation
 - Style elements
 	- Form elements
 	- Responsive tables
@@ -33,7 +34,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 2.0.1 - 2018-11-XX - XXXXXX
 
+### Added
+- Add hamburger menu icon
+
 ### Changed
+- Change navigation button markup
 - Style button element
 
 ### Removed